push que esqueci, mals pedro

+++ src/avaliacao/avaliacao.service.ts
@@ -9,25 +9,31 @@ export class AvaliacaoService {
 
   async create(data: CreateAvaliacaoDto) {
     // Verifica se o usuário existe
-    const user = await this.prisma.user.findUnique({
+    const userExists = await this.prisma.user.findUnique({
       where: { id: data.userId },
     });
-    if (!user) throw new NotFoundException('Usuário não encontrado');
+    if (!userExists) {
+      throw new NotFoundException('Usuário não encontrado');
+    }
 
     // Verifica se o professor existe
-    const professor = await this.prisma.professor.findUnique({
+    const professorExists = await this.prisma.professor.findUnique({
       where: { id: data.professorId },
     });
-    if (!professor) throw new NotFoundException('Professor não encontrado');
+    if (!professorExists) {
+      throw new NotFoundException('Professor não encontrado');
+    }
 
     // Verifica se o curso existe
-    const course = await this.prisma.courses.findUnique({
+    const courseExists = await this.prisma.courses.findUnique({
       where: { id: data.courseId },
     });
-    if (!course) throw new NotFoundException('Curso não encontrado');
+    if (!courseExists) {
+      throw new NotFoundException('Curso não encontrado');
+    }
 
     // Cria a avaliação após as validações
-    return await this.prisma.avaliacao.create({
+    return this.prisma.avaliacao.create({
       data: {
         text: data.text,
         userId: data.userId,
@@ -41,7 +47,7 @@ export class AvaliacaoService {
   }
 
   async findAll() {
-    return await this.prisma.avaliacao.findMany({
+    return this.prisma.avaliacao.findMany({
       include: {
         user: true,
         professor: true,
@@ -57,20 +63,42 @@ export class AvaliacaoService {
         user: true,
         professor: true,
         course: true,
+        comments: {
+          include: {
+            user: true,
+          },
+        },
       },
     });
-    if (!avaliacao) throw new NotFoundException('Avaliação não encontrada');
+
+    if (!avaliacao) {
+      throw new NotFoundException('Avaliação não encontrada');
+    }
     return avaliacao;
   }
 
   async deleteAvaliacao(id: number) {
-    return await this.prisma.avaliacao.delete({
+    const avaliacaoExists = await this.prisma.avaliacao.findUnique({
+      where: { id },
+    });
+    if (!avaliacaoExists) {
+      throw new NotFoundException('Avaliação não encontrada');
+    }
+
+    return this.prisma.avaliacao.delete({
       where: { id },
     });
   }
 
   async updateAvaliacao(id: number, data: UpdateAvaliacaoDto) {
-    return await this.prisma.avaliacao.update({
+    const avaliacaoExists = await this.prisma.avaliacao.findUnique({
+      where: { id },
+    });
+    if (!avaliacaoExists) {
+      throw new NotFoundException('Avaliação não encontrada');
+    }
+
+    return this.prisma.avaliacao.update({
       where: { id },
       data,
     });
